Add files via upload
Hello, I created a speed typing game! You are given a certain amount of seconds to type in the displayed word. If you do not complete typing the displayed word the game ends! While you play your score is shown below! This was created using mainly JavaScript, HTML, and CSS.